DE Competition updates pt.2
Moved some defines from Auto2015.c to Auto2015.h
Added functionality to lift function to fail if it can't raise
We also replaced the right IR servo, so the values had to change.
We redid the logic to determine the position of the center turnstile
more accurately, using both IR seekers.
I also fixed Auto2015ramp.c, but we haven't tested it.
I made a new program, Diag.c, that is for battery utilities. It reads
the NXT & external batteries and determines their statuses, but I
haven't been able to test the values. It also displays the status of the
SMUX.
We made a new program (IRChart) to temporarily replace IRGraph  because
IRGraph was freezing the NXT and wasn't operational. We modified IRGraph
to work with the SMUX and 2 IR beacons, but it isn't finished.
Changed some values in libAuto
Commented some stuff
Updated ServoControl.c